[MOV] mail,web_editor: move test_font_to_img to mail
This was missed in https://github.com/odoo/odoo/pull/224304.
