+++ addons/mail/tests/__init__.py
@@ -2,6 +2,7 @@
 
 from . import test_discuss_tools
 from . import test_fetchmail
+from . import test_font_to_img
 from . import test_ir_mail_server
 from . import test_ir_ui_menu
 from . import test_ir_websocket
